Inferfacify TblPropsToggleRegistry for internal extension
